Added Timber and Stetho to the Android App.
Completed the following steps:
Step 0: Starting Point
This is the starting point for the final project, which is provided to you in the course repository.
It contains an activity with a banner ad and a button that purports to tell a joke, but actually just complains. The banner ad was set up following the instructions here.
Step 1: Create a Java library
Your first task is to create a Java library that provides jokes. Create a new Gradle Java project either using the Android Studio wizard, or by hand. Then introduce a project dependency between your app and the new Java Library. If you need review, check out demo 4.01 from the course code.
Make the button display a toast showing a joke retrieved from your Java joke telling library.
Step 2: Create an Android Library
Create an Android Library containing an Activity that will display a joke passed to it as an intent extra. Wire up project dependencies so that the button can now pass the joke from the Java Library to the Android Library.